+++ src/main/java/com/booking/db/repository/ConfirmLinkRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ConfirmLinkRepository extends CrudRepository<ConfirmLink, Long> {
+public interface ConfirmLinkRepository extends CrudRepository<ConfirmLink, Integer> {
     List<ConfirmLink> findByUserEmail(String email);
 }
